% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/communes_ch_indicators.R
\name{loadCommunesCHportraits}
\alias{loadCommunesCHportraits}
\title{Swiss communes socio-demographic indicators}
\usage{
loadCommunesCHportraits()
}
\value{
a matrix
}
\description{
Load socio-demographic data by the Swiss Statistical Office.
}
\details{
The matrix has the commune BFS code as rowname and 3 attributes: \code{communeName} the text name of the commune (same length as the matrix length),  \code{indicatorYear} & \code{indicatorGroup} respectively the year and the category of the communal indicator (both of same length as the matrix width). See \url{http://www.bfs.admin.ch/bfs/portal/fr/index/regionen/02/key.html}
}
\examples{
communeData <- loadCommunesCHportraits()
colnames(communeData)
rownames(communeData)

# Select only "surface" indicators
colIdx <- which(attr(communeData, "indicatorGroup") == "Surface")
head(communeData[,colIdx])

# Get all the commune names
communesCH <- loadCommunesCHdata(start = "2012-01-01", end = Sys.Date())
}
\seealso{
Office fédéral de la statistique > Les Régions > Communes > Données et explications (portraits): http://www.bfs.admin.ch/bfs/portal/fr/index/regionen/02/daten.html
}

